requirements.txt
asgiref
Django
djangorestframework
psycopg2-binary
Pygments
pytz
sqlparse
gunicorn